Disable scroll-to-top for menu. Closes #186
Scroll-to-top does not work in case there are multiple view controllers - for us this is the case when the user is signed in and the menu is shown besides (or below) the main view controller.

From the Apple docs:
This gesture works on a single visible scroll view; if there are multiple scroll views (for example, a date picker) with this property set, or if the delegate returns NO in scrollViewShouldScrollToTop:, UIScrollView ignores the request.