Cleaned up GHOrganization. Fixes #322
Check the values before setting them, because the organizations list does not include all fields. This unsets some fields when reloading the orgs, after an org has already been fully loaded (because the orgs are cached).

This commit also removes unused attributes from GHOrganization. These field are present in the API response, but you cannot fill the in on GitHub.com and they also aren't shown there in the organization interface.