Disable "Mark all as read" in some cases. Closes #278
Right now the GitHub API does not fully support pagination for notifications. This is a temporary workaround that disables the "Mark all as read" button if the maximum number of 50 notifications was loaded, because in this case we do not know whether or not there are more notifications (which the user did not see) - so don't make it possible to mark all as read in this case.