Load cached gravatar via NSData. Closes #347
The problem was a race condition that occured while loading the cached gravatar while at the same time the avatar cache got cleared. UIImage. imageWithData loads the image as a valid image (instead of returning nil) in case it can read the header. The image data was corrupted when the image got deleted while loading it, that's why only a black gravatar was shown.

For more details, see this answer on SO:
http://stackoverflow.com/a/10685962/183537